refactor(reporting): extract shared report staging helpers (#239)
